Added more styling and profile

+++ src/main/java/com/codeup/springblog/controllers/PostController.java
@@ -1,8 +1,10 @@
 package com.codeup.springblog.controllers;
 
 import com.codeup.springblog.models.Post;
+import com.codeup.springblog.models.User;
 import com.codeup.springblog.repos.PostRepository;
 import com.codeup.springblog.repos.UserRepository;
+import com.codeup.springblog.services.AuthenticationService;
 import com.codeup.springblog.services.EmailService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -16,8 +18,10 @@ public class PostController {
     private PostRepository postDao;
     private UserRepository userDao;
     private final EmailService emailService;
+    private AuthenticationService authSvc;
 
-    public PostController(PostRepository postDao, UserRepository userDao, EmailService emailService) {
+    public PostController(AuthenticationService authSvc,PostRepository postDao, UserRepository userDao, EmailService emailService) {
+        this.authSvc = authSvc;
         this.emailService = emailService;
         this.postDao = postDao;
         this.userDao = userDao;
@@ -49,15 +53,19 @@ public class PostController {
 
     @PostMapping("/posts/create")
     public String insert(@ModelAttribute Post post) {
-        post.setOwner(userDao.findOne(1L));
+        post.setOwner(userDao.findOne(((User)authSvc.getCurUser()).getId()));
         postDao.save(post);
-        emailService.prepareAndSend(post);
+//        emailService.prepareAndSend(post);
         return "redirect:/posts";
     }
 
     @GetMapping("/posts/{id}/edit")
     public String edit(@PathVariable long id, Model model) {
-        model.addAttribute("post", postDao.findOne(id));
+        Post post = postDao.findById(id);
+        if(((User)authSvc.getCurUser()).getId() != post.getOwner().getId()) {
+            return "redirect:/posts";
+        }
+        model.addAttribute("post", post);
         model.addAttribute("action", "/posts/"+id+"/edit");
         model.addAttribute("title", "Edit Post");
 
@@ -66,13 +74,19 @@ public class PostController {
 
     @PostMapping("/posts/{id}/edit")
     public String update(@ModelAttribute Post post) {
-        postDao.save(post);
+        post.setOwner(postDao.findById(post.getId()).getOwner());
+        if(((User)authSvc.getCurUser()).getId() == post.getOwner().getId()) {
+            postDao.save(post);
+        }
         return "redirect:/posts";
     }
 
     @PostMapping("/posts/{id}/delete")
     public String delete(@PathVariable long id) {
-        postDao.delete(id);
+        Post post = postDao.findById(id);
+        if(((User)authSvc.getCurUser()).getId() == post.getOwner().getId()) {
+            postDao.delete(id);
+        }
         return "redirect:/posts";
     }
